backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "datetime>=5.5",
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "motor>=3.7.1",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.0.0",
    "pip>=25.1.1",
    "numpy>=2.3.0",
    "langgraph>=0.2.38",
    "bs4>=0.0.2",
    "gitingest>=0.3.1",
    "langchain-tavily>=0.2.11",
    "tavily-python>=0.7.12",
    "pymupdf>=1.26.5",
    "pymupdf4llm>=0.0.27",
    "google-genai>=1.46.0",
    "sse-starlette>=2.2.1",
    "httpx>=0.28.1",
    "langchain-openai>=0.3.35",
    "langchain-anthropic>=0.3.22",
    "langchain-ollama>=0.3.10",
    "cryptography>=46.0.5",
]

[dependency-groups]
dev = [
    "uvicorn>=0.34.3",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl" }
